backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic-settings>=2.3,<3.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
yfinance>=0.2.54,<1.0
openai>=1.35,<2.0
pytest>=8.2,<9.0
